backend/requirements.txt
fastapi
uvicorn[standard]
prisma
scrapy
pydantic
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
email-validator
apscheduler
asyncpg
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client